requirements.txt
streamlit==1.24.0
anthropic==0.3.0
sentence-transformers==2.2.2
torch==2.0.1
python-dotenv